move pkg/kubelet/server to cmd/kubelet/app
