+++ cmd/kubelet/app/plugins.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package server
+package app
 
 // This file exists to force the desired plugin implementations to be linked.
 import (
